+++ cuda/src/kernels/conv.rs
@@ -50,13 +50,13 @@ impl ConvKernel for Generic {
         let input = get_cuda_view(input);
 
         launcher.set_view(&input);
-        launcher.set_el::<i64>(*input_shape.n().unwrap_or(&1));
-        launcher.set_el::<i64>(*input_shape.c());
-        launcher.set_slice::<i64>(input_shape.hw_dims());
+        launcher.set_el::<i32>(*input_shape.n().unwrap_or(&1));
+        launcher.set_el::<i32>(*input_shape.c());
+        launcher.set_slice::<i32>(input_shape.hw_dims());
 
-        launcher.set_el::<i64>(*input_shape.n_stride().unwrap_or(&0));
-        launcher.set_el::<i64>(*input_shape.c_stride());
-        launcher.set_slice::<i64>(input_shape.hw_strides());
+        launcher.set_el::<i32>(*input_shape.n_stride().unwrap_or(&0));
+        launcher.set_el::<i32>(*input_shape.c_stride());
+        launcher.set_slice::<i32>(input_shape.hw_strides());
 
         let kfmt = op.kernel_fmt;
         let co_per_group = op.pool_spec.output_channels / op.group;
@@ -65,17 +65,17 @@ impl ConvKernel for Generic {
         let weights_view = get_cuda_view(weights);
         launcher.set_view(&weights_view);
         // split go_i_h_w in g_o_i_h_w
-        launcher.set_el::<i64>(op.group);
-        launcher.set_el::<i64>(co_per_group);
-        launcher.set_slice::<i64>(&weights.shape()[1..]);
+        launcher.set_el::<i32>(op.group);
+        launcher.set_el::<i32>(co_per_group);
+        launcher.set_slice::<i32>(&weights.shape()[1..]);
 
         let group_stride = weights.strides()[0] as usize * co_per_group;
-        launcher.set_el::<i64>(group_stride);
-        launcher.set_slice::<i64>(weights.strides());
+        launcher.set_el::<i32>(group_stride);
+        launcher.set_slice::<i32>(weights.strides());
 
         let bias_view = get_cuda_view(bias);
         launcher.set_view(&bias_view);
-        launcher.set_el::<i64>(if bias.rank() == 0 {
+        launcher.set_el::<i32>(if bias.rank() == 0 {
             0usize // scalar bias: stride = 0 is broadcasting
         } else {
             1usize
@@ -83,14 +83,14 @@ impl ConvKernel for Generic {
 
         let padding = op.pool_spec.computed_padding(input_shape.hw_dims());
         for d in 0..input_shape.hw_rank() {
-            launcher.set_el::<i64>(padding[d].pad_before);
+            launcher.set_el::<i32>(padding[d].pad_before);
         }
 
         let strides = op.pool_spec.strides();
-        launcher.set_slice::<i64>(&strides);
+        launcher.set_slice::<i32>(&strides);
 
         let dilations = op.pool_spec.dilations();
-        launcher.set_slice::<i64>(&dilations);
+        launcher.set_slice::<i32>(&dilations);
 
         dbg!(output.shape());
 
@@ -98,13 +98,13 @@ impl ConvKernel for Generic {
         dbg!(&output_shape);
         let output = get_cuda_view(output);
         launcher.set_view(&output);
-        launcher.set_el::<i64>(*output_shape.n().unwrap_or(&1));
-        launcher.set_el::<i64>(*output_shape.c());
-        launcher.set_slice::<i64>(output_shape.hw_dims());
+        launcher.set_el::<i32>(*output_shape.n().unwrap_or(&1));
+        launcher.set_el::<i32>(*output_shape.c());
+        launcher.set_slice::<i32>(output_shape.hw_dims());
 
-        launcher.set_el::<i64>(*output_shape.n_stride().unwrap_or(&0));
-        launcher.set_el::<i64>(*output_shape.c_stride());
-        launcher.set_slice::<i64>(output_shape.hw_strides());
+        launcher.set_el::<i32>(*output_shape.n_stride().unwrap_or(&0));
+        launcher.set_el::<i32>(*output_shape.c_stride());
+        launcher.set_slice::<i32>(output_shape.hw_strides());
 
         let cfg = LaunchConfig {
             grid_dim: (
